fix start and end positions of match
the `match.pos` and `match.endpos` are not the start and end of the matched
group but the start and end arguments passed to the `re.match()` method.

use `match.start()` and `match.end()` to find the beginning and the end of
the matched group.

see https://docs.python.org/3/library/re.html#match-objects