Add maubot server alias feature
This means you can add an alias for a maubot server when logging in,
and then use that alias instead of typing the full URL each time.

Mostly useful for people with more than one maubot instance, as the
default server already didn't need to be typed out each time.