Don't update displayname/avatar when unset
Avoid sending no-op updates to the homeserver when there is no
displayname or avatar_url set. The database stores these as empty
strings, but the homeserver returns them as None, so when compared they
appear different and permit an update.

This is a workaround that should fix https://github.com/maubot/maubot/issues/235
by coercing empty strings that are stored in the database to/from
null/None values during saves/loads.

Signed-off-by: Joe Groocock <me@frebib.net>